[#23] Rename seeder & update folder structure
